Addition of more unit tests, move the emailsenderserver tests to a different folder
