+++ src/backend/TrafficCourts/Test/TrafficCourts.Test.csproj
@@ -31,6 +31,7 @@
 
   <ItemGroup>
     <Folder Include="Messaging\" />
+    <Folder Include="Workflow.Service\Features\Mail\" />
   </ItemGroup>
 
   <ItemGroup>
