+++ src/backend/TrafficCourts/Workflow.Service/appsettings.json
@@ -23,7 +23,7 @@
   },
   "EmailConfiguration": {
     "Sender": "DoNotReply@gov.bc.ca",
-    "AllowList": ["*@gov.bc.ca", "*@nttdata.com"]
+    "AllowList": ["@gov.bc.ca", "@nttdata.com"]
   },
   "RetryConfiguration": {
     "RetryTimes": 5,
